+++ delivery-service/src/main/java/com/eighteenthstreet/deliveryservice/domain/exception/DeliveryNotFoundException.java
@@ -1,10 +1,10 @@
 package com.eighteenthstreet.deliveryservice.domain.exception;
 
-import com.eighteenthstreet.deliveryservice.presentation.exception.ErrorCode;
 import com.eighteenthstreet.deliveryservice.presentation.exception.error.CustomException;
+import exception.ErrorCode;
 
 public class DeliveryNotFoundException extends CustomException {
-	public DeliveryNotFoundException(ErrorCode code) {
-		super(code);
-	}
+    public DeliveryNotFoundException(ErrorCode code) {
+        super(code);
+    }
 }
