+++ delivery-service/src/main/java/com/eighteenthstreet/deliveryservice/domain/exception/InvalidDeliveryException.java
@@ -1,10 +1,10 @@
 package com.eighteenthstreet.deliveryservice.domain.exception;
 
-import com.eighteenthstreet.deliveryservice.presentation.exception.ErrorCode;
 import com.eighteenthstreet.deliveryservice.presentation.exception.error.CustomException;
+import exception.ErrorCode;
 
 public class InvalidDeliveryException extends CustomException {
-	public InvalidDeliveryException(ErrorCode code) {
-		super(code);
-	}
+    public InvalidDeliveryException(ErrorCode code) {
+        super(code);
+    }
 }
