+++ delivery-service/src/main/java/com/eighteenthstreet/deliveryservice/domain/model/Delivery.java
@@ -3,8 +3,8 @@ package com.eighteenthstreet.deliveryservice.domain.model;
 
 import base.BaseEntity;
 import com.eighteenthstreet.deliveryservice.domain.exception.InvalidDeliveryException;
-import com.eighteenthstreet.deliveryservice.presentation.exception.ErrorCode;
 import com.eighteenthstreet.deliveryservice.presentation.request.CreateDeliveryRequest;
+import exception.ErrorCode;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.SQLRestriction;
