--- delivery-service/src/main/java/com/eighteenthstreet/deliveryservice/presentation/exception/ErrorResponse.java
@@ -1,19 +1,0 @@
-package com.eighteenthstreet.deliveryservice.presentation.exception;
-
-public class ErrorResponse {
-	private final String code;
-	private final String message;
-
-	public ErrorResponse(ErrorCode errorCode) {
-		this.code = errorCode.getCode();
-		this.message = errorCode.getMessage();
-	}
-
-	public String getCode() {
-		return code;
-	}
-
-	public String getMessage() {
-		return message;
-	}
-}
