+++ delivery-service/src/main/java/com/eighteenthstreet/deliveryservice/presentation/exception/GlobalExceptionHandler.java
@@ -1,18 +1,18 @@
 package com.eighteenthstreet.deliveryservice.presentation.exception;
 
+import com.eighteenthstreet.deliveryservice.presentation.exception.error.CustomException;
+import exception.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import com.eighteenthstreet.deliveryservice.presentation.exception.error.CustomException;
-
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
-	@ExceptionHandler(CustomException.class)
-	public ResponseEntity<ErrorResponse> handleCustomException(CustomException e) {
-		return ResponseEntity.status(e.getErrorCode().getStatus())
-			.body(new ErrorResponse(e.getErrorCode()));
-	}
+    @ExceptionHandler(CustomException.class)
+    public ResponseEntity<ErrorResponse> handleCustomException(CustomException e) {
+        return ResponseEntity.status(e.getErrorCode().getStatus())
+                .body(new ErrorResponse(e.getErrorCode()));
+    }
 
 }
