+++ delivery-service/src/main/java/com/eighteenthstreet/deliveryservice/presentation/exception/error/CustomException.java
@@ -1,15 +1,14 @@
 package com.eighteenthstreet.deliveryservice.presentation.exception.error;
 
-import com.eighteenthstreet.deliveryservice.presentation.exception.ErrorCode;
-
+import exception.ErrorCode;
 import lombok.Getter;
 
 @Getter
 public class CustomException extends RuntimeException {
-	private final ErrorCode errorCode;
+    private final ErrorCode errorCode;
 
-	public CustomException(ErrorCode errorCode) {
-		super(errorCode.getMessage());
-		this.errorCode = errorCode;
-	}
+    public CustomException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
 }
